chore: 어노테이션 수정, authService를 userService로 병합
