+++ src/main/java/project/Ecommerce/controller/UserController.java
@@ -11,7 +11,6 @@ import org.springframework.web.bind.annotation.RestController;
 import project.Ecommerce.dto.ReIssue;
 import project.Ecommerce.dto.SignIn;
 import project.Ecommerce.dto.SignUp;
-import project.Ecommerce.service.AuthService;
 import project.Ecommerce.service.UserService;
 
 @Slf4j
@@ -20,7 +19,6 @@ import project.Ecommerce.service.UserService;
 @RequestMapping("/user")
 public class UserController {
   private final UserService userService;
-  private final AuthService authService;
 
   @PostMapping("/signup")
   public ResponseEntity<SignUp.Response> signUp(@RequestBody @Valid SignUp.Request request) {
@@ -34,6 +32,6 @@ public class UserController {
 
   @PostMapping("/reissue")
   public ResponseEntity<ReIssue.Response> reIssue(@RequestBody @Valid ReIssue.Request request) {
-    return ResponseEntity.ok(authService.reIssue(request));
+    return ResponseEntity.ok(userService.reIssue(request));
   }
 }
