+++ src/main/java/project/Ecommerce/security/GetAuthentication.java
@@ -4,14 +4,14 @@ import io.jsonwebtoken.Jwts;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.annotation.ComponentScan;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.stereotype.Component;
 import project.Ecommerce.service.MemberDetailService;
 
 @Slf4j
-@Component
+@ComponentScan
 @RequiredArgsConstructor
 public class GetAuthentication {
