--- src/main/java/project/Ecommerce/service/AuthService.java
@@ -1,8 +1,0 @@
-package project.Ecommerce.service;
-
-import project.Ecommerce.dto.ReIssue;
-
-public interface AuthService {
-
-  ReIssue.Response reIssue(ReIssue.Request request);
-}
