+++ src/main/java/project/Ecommerce/service/UserService.java
@@ -1,5 +1,6 @@
 package project.Ecommerce.service;
 
+import project.Ecommerce.dto.ReIssue;
 import project.Ecommerce.dto.SignIn;
 import project.Ecommerce.dto.SignUp;
 
@@ -7,5 +8,7 @@ public interface UserService {
 
   SignUp.Response signUp(SignUp.Request request);
   SignIn.Response signIn(SignIn.Request request);
+  ReIssue.Response reIssue(ReIssue.Request request);
+
 
 }
